Worked on adding the mermaid pickup of the kid. Added controllers to mermaid and kid
Froze the Z rotation for both characters, that way the characters won't rotate.
Add project settings for input for the "Switch" key = "G" and "Activate" key = "E". That way we could use inputs from other controllers or the user wishes to chage the controls if need be.
Updated the switching code with the new project settings for input.

In the switching, I removed the code where the position of the mermaid would go to the kid and vice versa. That way the mermaid can go somewhere else then the kid.